command: Pack struct in big-endian byte order
It's already done in all other integer types and was probably forgotten
here. It's not a big issue because this path is only called for sending
mandatory parameters, and all integer mandatory parameters I could find
are only 1 byte long. However, add it for completeness and to avoid any
future issues in case of refactor.